style(core,tui): replace &PathBuf with &Path and apply idiomatic Rust lints
Migrate all Tool trait method signatures and internal helpers from
&PathBuf to &Path for broader API generality. Fix numerous Clippy
warnings across the codebase including strip_prefix usage over manual
string slicing, derive(Default) over hand-rolled impls, vec![] over
push loops, iterator adapters (find, contains, clamp, div_ceil,
repeat_n), and remove unnecessary allocations and unused imports.
Also add bidirectional Orchestrator↔DynamicOrchestrator conversions
and await channel sends in the TUI event loop.

+++ radiumical-tui/src/settings.rs
@@ -585,7 +585,7 @@ impl SettingsBoard {
             .map(|i| i.label.chars().count())
             .max()
             .unwrap_or(12) as u16;
-        let max_value = (r.width.saturating_sub(label_w + 8) as u16).max(10) as usize;
+        let max_value = (r.width.saturating_sub(label_w + 8)).max(10) as usize;
         let block = Block::default()
             .borders(Borders::ALL)
             .border_type(BorderType::Rounded)
